Ignore zoomed-out maps.
For some weird reason, either because the maps are in a
non-standard world, or because of the image-on-map plugin
(have not investigated the cause), zoomed out maps record
their centre coordinates at an incorrect location. They also
record the `scale` as 4 when it should be 1 for the first
scale level.

So to avoid spurious markers, we simply pretend these map
NBT data files don't exist when we see a non-zero scale.