Ignore map IDs that are not in an item frame.
Map IDs that have not been "blessed" by being placed in an
item frame will not be counted as "existing" for the purposes
of assigning dynmap markers.

The `frames` list in the map NBT file does not contain the
world name; it only has a list of item frame entity IDs,
x/Y/Z coordinates and item rotation values.